+++ src/MyLab.Search.Searcher/Controllers/TokenControllerV4.cs
@@ -7,21 +7,21 @@ using MyLab.WebErrors;
 namespace MyLab.Search.Searcher.Controllers
 {
     [ApiController]
-    [Route("v1/token")]
-    public class TokenControllerV1 : ControllerBase
+    [Route("v4/token")]
+    public class TokenControllerV4 : ControllerBase
     {
         private readonly ITokenService _tokenService;
 
-        public TokenControllerV1(ITokenService tokenService)
+        public TokenControllerV4(ITokenService tokenService)
         {
             _tokenService = tokenService;
         }
 
         [HttpPost]
         [ErrorToResponse(typeof(TokenizingDisabledException), HttpStatusCode.Locked)]
-        public IActionResult Post([FromBody] TokenRequestV1 tokenRequest)
+        public IActionResult Post([FromBody] TokenRequestV4 tokenRequest)
         {
-            return Ok(_tokenService.CreateSearchToken(tokenRequest.ToV3()));
+            return Ok(_tokenService.CreateSearchToken(tokenRequest));
         }
     }
 }