+++ src/MyLab.Search.Searcher/Models/ClientSearchRequestV4.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using Newtonsoft.Json;
+﻿using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 
 #if IS_CLIENT
@@ -8,15 +7,14 @@ namespace MyLab.Search.Searcher.Client
 namespace MyLab.Search.Searcher.Models
 #endif
 {
-    [Obsolete]
-    public class ClientSearchRequestV2
+    public class ClientSearchRequestV4
     {
         [JsonProperty("query")]
         public string Query { get; set; }
         [JsonProperty("filters")]
         public FilterRef[] Filters { get; set; }
         [JsonProperty("sort")]
-        public string Sort { get; set; }
+        public SortingRef Sort { get; set; }
         [JsonProperty("offset")]
         public int Offset { get; set; }
         [JsonProperty("limit")]
@@ -24,20 +22,5 @@ namespace MyLab.Search.Searcher.Models
         [JsonProperty("queryMode")]
         [JsonConverter(typeof(StringEnumConverter))]
         public QuerySearchStrategy QuerySearchStrategy { get; set; }
-
-        public ClientSearchRequestV3 ToV3()
-        {
-            return new ClientSearchRequestV3
-            {
-                Query = Query,
-                QuerySearchStrategy = QuerySearchStrategy,
-                Offset = Offset,
-                Sort = Sort != null 
-                    ? new SortingRef { Id = Sort }
-                    : null,
-                Limit = Limit,
-                Filters = Filters
-            };
-        }
     }
 }