+++ src/MyLab.Search.Searcher/Services/ITokenService.cs
@@ -5,7 +5,7 @@ namespace MyLab.Search.Searcher.Services
     public interface ITokenService
     {
         bool IsEnabled();
-        string CreateSearchToken(TokenRequestV3 request);
-        NamespaceSettingsV3 ValidateAndExtractSettings(string token, string ns);
+        string CreateSearchToken(TokenRequestV4 request);
+        IndexSettingsV4 ValidateAndExtractSettings(string token, string idxId);
     }
 }